MCSSA(simsalabim)	R Documentation

Monte Carlo SSA (MCSSA)

Description

Simple implementation of MCSSA

Usage


MCSSA(dSSA, ...)
## S3 method for class 'decompSSA':
MCSSA(dSSA, x, n, conf = 0.95,
      keepSurr = FALSE,ar.method="mle", ...)


Arguments

dSSA 	 Object of class decompSSA.

x 	 A numeric vector, interpreted as time series.

n 	 Number of surrogates to be computed.

conf 	 Confidence limit.

keepSurr 	 Whether to keep the surrogate estimates of lambda or not.

ar.method 	 Method to estimate AR1 parameters, passed on to ar. One of "yule-walker", "burg", "ols", "mle", "yw".

... 	 Arguments passed to methods.

Details

A simple implementation of the MCSSA, testing against the AR1 hypothesis. The parameters of the AR1 process are estimated via ar .

The confidence level conf can be specified either as number < 1 or as vector of length two with sum(conf)==1.

Value

An object of class MCSSA

lambda 	 The eigenvalues, ordered by decreasing frequency.

freq 	 Dominant frequency of the eigenvectors, ordered by decreasing value.

rank 	 Rank of the eigenvalues, ordered by decreasing frequency.

ar1 	 The noise model, output of ar.mle

upper 	 Upper confidence limit, ordered by decreasing frequency.

lower 	 Lower confidence limit, ordered by decreasing frequency.

conf 	 Vector of length two, specifying the upper and lower confidence bounds of the test.

N 	 Length of the initial series.

L 	 The embedding dimension.

call 	 Call of the generating function.

surLambda 	 A matrix containing the estimates of the surrogate eigenvalues, ordered by the dominant frequency in its columns. Only available if keepSurr=TRUE.

Warning

Large computational demands

Author(s)

Lukas Gudmundsson

References

Allen, M. & Smith, L. Monte Carlo SSA: Detecting irregular oscillations in the Presence of Colored Noise. Journal of Climate, 1996, 9, 3373-3404

See Also

sdTest, decompSSA, plot.MCSSA

Examples


n <- arima.sim(n=200,alist(ar=0.4))
n <- as.vector(scale(n))
x <- sin(seq(0,15*pi,len=200))
xn <- x+n/2
xn.dc <- decompSSA(xn,L=40)
xn.mct <- MCSSA(xn.dc,xn,n=50)
plot(xn.mct)

[Package simsalabim version 0.1-1 Index]