decompSSAM(simsalabim)	R Documentation

Singular System Analysis for time series with missing values

Description

Implementation of the SSA algorithm for time series with missing values following Golyandina and Osipov (2007).

Usage


decompSSAM(x, L, tau = 0, toeplitz = FALSE, getFreq = TRUE)
reconSSAM(dSSAM, x, groups,
      method = list(alpha = "PI", beta = "simultaneous"))


Arguments

x 	 A vector representing the time series.

L 	 Embedding dimension.

tau 	 A number ranging from 0 to 1, indicating the maximal proportion of missing values within columns of the trajectory matrix for the usage of a modification of the inner product. If tau is negative it is treated as if it is zero. Only applies if toeplitz = FALSE.

toeplitz 	 Whether to use the Toeplitz modification of SSA for stationary time series or not.

getFreq 	 Whether dominant frequencies of the eigenvectors shall be determined.

dSSAM 	 Output of decompSSAM

groups 	 A list of vectors. Each vector is representing a selection of eigenvalues and eigenvectors which shall be used to compute reconstructed components.

method 	 Method or combination of methods that shall be used to estimate the reconstructed components at the position of missing values. Currently one of "PC" or
list(alpha="PI",beta="simultaneous").

Details

decompSSAM performs the SSA decomposition whereas reconSSAM performs the SSA reconstruction, which is a type of band pass filtering.

In general the application of SSA for time series with missing values follows the same principals as for standard SSA. For general comments on the application of these functions see the documentation of decompSSA.

SSA embeds lagged copies of a time series x into a augmented matrix X (trajectory matrix ). In a second step the orthonormal basis of X is found via singular value decomposition (SVD). One of the internal steps of SVD is the computation of XX^T. There are two strategies available to obtain this product if values of x are missing. The first possibility is to omit any column of X and to compute XX^T for that reduced matrix. The other possibility is to use a modification of the inner product for vectors containing missing values. The threshold parameter tau controls the computation of XX^T. It gives the acceptable proportion of missing values whithin a vector for the applicatoin of the modified inner product. If tau is <=0, all columns of X containing any missing value are ommited. If tau >=1 no column of X is ommited. For any value of tau in between all columns of X having a propotion of more or equal than tau are omitted. There are several possibilities to capture the reconstructed components. One is the recovery by means of principal component, which has first been introduced by Schoellhamer (2001). Golyandina and Osipov split the recovery of the reconstructed signals into two steps alpha and beta (see reference for more details). Currently only one method for each step is implemented, namely alpha="PI", denoting the usage of the Pi - projector and beta="simultaneous", denoting simultaneous filling in.

Value

The output of decompSSAM is an object of class decompSSAM inheriting from decompSSA with following items:

lambda 	 The eigenvalues, ordered decreasing.

U 	 The eigenvectors (columns), ordered by decreasing eigenvalues.

freq 	 Dominant frequency of the eigenvectors, ordered by decreasing eigenvalues.

rank 	 Rank of the eigenvalues, ordered by decreasing eigenvalues.

N 	 Length of the input series.

L 	 Embedding dimension or window length.

toeplitz 	 Logical, indicates if Toeplitz modification has been used.

numMisssing 	 The number of missing entries in the input.

tau 	 The user specification of tau.

seriesName 	 Name of input series.

call 	 Call of generating function.

The output of reconSSAM is a matrix with length(groups) column and length(x) rows. Each columns represents the sum of the reconstructed components defined by the list entries of groups.

Warning

May cause extreme memory demands. reconSSAM is computionally expensive.

Author(s)

Lukas Gudmundsson

References

Golyandina, N. & Osipov, E. The "Caterpillar"-SSA method for analysis of time series with missing values. Journal of Statistical Planning and Inference, 5th St. Petersburg Workshop on Simulation, 2007, 137, 2642-2653 http://www.gistatgroup.com/cat/mvssa1en.pdf

Schoellhamer, D. Singular spectrum analysis for time series with missing data. Geophysical Research Letters, 2001, 28, 3187-3190

See Also

plot.decompSSAM, decompSSA, sdTest

Examples


x <- sin(seq(0,10*pi,len=200))
x <- x + rnorm(x)/2
x[100:120] <- NA
x.dc <- decompSSAM(x,L=40)
plot(x.dc,by="rank",log="")
x.rc1 <- reconSSAM(x.dc, x, groups=list(1:2),
  method=list(alpha = "PI", beta = "simultaneous"))
x.rc2 <- reconSSAM(x.dc, x, groups=list(1:2),
  method="PC")
# compare result with input signal
plot(x,type="l")
lines(x.rc1,col="red",lwd=2,lty=1)
lines(x.rc2,col="black",lwd=2, lty=2)
lines(sin(seq(0,10*pi,len=200)),col="blue",lwd=2)

[Package simsalabim version 0.1-1 Index]