kgSSAM(simsalabim)	R Documentation

Itterative filling in of missing values in time series based on SSA

Description

Implementation of the iterative SSA algorithm for time series with missing values introduced by Kondrashov & Ghil (2006).

Usage


kgSSAM(x, L, toeplitz = FALSE, K, max.inner = 10,
       eps = 0.98, getFreq = TRUE)
CV.kgSSAM(x, L, K_min = 1, K_max = 10, cv.gap = 0.05, runs = 30, ...)


Arguments

x 	 A vector representing the time series.

L 	 Embedding dimension or window length.

toeplitz 	 Use toeplitz variant of SSA for stationary time series.

getFreq 	 Whether dominant frequencies of the eigenelements shall be determined.

K 	 Vector with index number (rank) of the eigenvectors to be used for the reconstruction. If it is a scalar the first K eigenvectors will be used.

max.inner 	 Maximum number of iterations for the estimation of each component represented by K. This is only used the error of reconstruction is not converging.

eps 	 Convergence criterion for the iteration which is estimating the reconstructed components specified by K. Currently this is implemented as the minimal coefficient of determination (R^2) value between the corresponding reconstructed components between two iteration steps.

K_min 	 The minimal number of reconstructed components to be used within cross validation.

K_max 	 The maximal number of reconstructed components to be used within cross validation.

cv.gap 	 Proportion of artificially introduced gaps for the cross validation procedure.

runs 	 Number of runs for the cross validation for each parameter combination.

... 	 arguments to be passed to kgSSAM.

Value

The output of kgSSAM is an object of class kgSSAM containing following items:

lambda 	 The eigenvalues, ordered by decreasing value.

U 	 The eigenvectors (columns), ordered by decreasing eigenvalues.

freq 	 Dominant frequency of the eigenvectors, ordered by decreasing eigenvalues.

rank 	 Rank of the eigenvalues, ordered by decreasing eigenvalues.

N 	 Length of the input series.

L 	 The embedding dimension.

toeplitz 	 Logical, indicates if toeplitz variant has been used.

x 	 The input series.

RC 	 The reconstructed components. Corresponding to K

R2 	 The value of the coefficient of determination (R^2) convergence criterion of the last iteration.

seriesName 	 Name of input series.

call 	 Call of the generating function

The output of CV.kgSSAM is a runs times (K_max - K_min) matrix containing the root mean square errors of the reconstructed components at the position of artificially introduced gaps for each cross validation run. Columns represent the runs. Rows represent the additive components used for the reconstruction.

Warning

Extreme computational demands.

Author(s)

Lukas Gudmundsson

References

Kondrashov, D. & Ghil, M. Spatio-temporal filling of missing points in geophysical data sets. Nonlinear Processes in Geophysics, 2006, 13, 151-159

See Also

decompSSAM and decompSSA

[Package simsalabim version 0.1-1 Index]