phaSyn(simsalabim)	R Documentation

Phase Synchronisation Analysis

Description

Computes the phase synchronisation index between to time series.

Usage


phaSyn(x1,x2,method="MRL",M=100)
phaSynMat(X,method="MRL",M=100,verbose=TRUE)

Arguments

x1, x2 	 Numeric vectors of the same length, interpreted as time series.

method 	 How to determine the phase synchronisation index. "MRL": mean resultant length. "SH": a measure based on the Shannon - Entropy.

M 	 Number of bins in the histogram of the cyclic phase difference. This is only used if method="SH" and is ignored otherwise.

X 	 A matrix containing time series in its columns.

verbose 	 If TRUE informations about the progress of the computation are printed.

Details

Phase synchronisation is quantified in two steps: First, the phase of both series is extracted by means of the Hilbert - transformation. Second, the distribuition of the cyclic phase difference of both series is checked for uniformity. In case of uniformity the series are not synchronous in respect to their phase.

Uniformity of the cyclic phase difference can be quantified in two ways. If method="MRL", the dispersion of the barycenter of the phase difference on the unit circle is calculated. In the case of method="SH", a histogram of the phase difference is built and its uniformity quantified by means of the Shannon - entropy.

The indices of both methods are scaled in such a manner that 1 means perfect synchronisation and 0 stands for no synchronisation.

Value

The output of phaSyn is an object of class pSyn containing:

rho 	 Phase synchronisation index.

phi 	 Cyclic phase difference.

call 	 The call of the generating function.

method 	 The method used to quantify peakedness of dsitribution.

name_x1 	 Name of series x1.

name_x2 	 Name of series x2.

The output of phaSynMat is a symmetric matrix containing the pairwise phase synchronisation index.

Author(s)

Lukas Gudmundsson

References

Allefeld, C. & Kurths, J. Testing for phase synchronization. International Journal of Bifurcation and Chaos, 2004, 14, 405-416

Rybski, D.; Havlin, S. & Bunde, A. Phase synchronization in temperature and precipitation records. Physica A: Statistical Mechanics and its Applications, Elsevier, 2003, 320, 601-610

See Also

synBoot

Examples


# phase synchronisation
x1 <- sin(1:100)
x2 <- cos(1:100)
phaSyn(x1,x2)

# phase synchronisation matrix
xx<-matrix(seq(0,6*pi,len=100),ncol=30,nrow=100)
colnames(xx) <- 1:30
xx[,1:10] <- sin(xx[,1:10])
xx[,11:20] <- sin(xx[,11:20]*2*pi)
xx[,21:30] <- sin(xx[,21:30]*4*pi)
cxx <- phaSynMat(xx,verbose=FALSE)


[Package simsalabim version 0.1-1 Index]