plot.MCSSA(simsalabim)	R Documentation

Plot Monte Carlo SSA

Description

Plots SSA - eigenspectrum and draws monte carlo error bars for each eigenvalue.

Usage


## S3 method for class 'MCSSA':
plot(x, by = "freq", normalize = FALSE, asFreq = TRUE,
      lam.pch = 19, lam.col = "black", lam.cex = 1, sig.col = "red",
      sig.pch = 19, sig.cex = 1, conf.col = "darkgray", log = "xy",
      ann = TRUE, legend = TRUE, axes = TRUE, ...)

Arguments

x 	 output of MCSSA.

by 	 How to order the eigenvalues. "freq": by frequency. "rank": by their rank (decreasingly ordered).

normalize 	 Normalise eigenvalues by their sum.

asFreq 	 If by="freq", shall the frequency be expressed as “frequency” or as “period”.

lam.pch 	 pch of the eigenvalues.

lam.col 	 col of the eigenvalues.

lam.cex 	 cex of the eigenvalues.

sig.col 	 col of the significant eigenvalues.

sig.pch 	 pch of the significant eigenvalues.

sig.cex 	 cex of the significant eigenvalues.

conf.col 	 col of the monte carlo confidence limits.

log 	 Use logarithmic axes, usage like log in par.

ann 	 Whether to do annotations or not.

legend 	 Whether to produce a legend or not.

axes 	 Whether to draw axes or not.

... 	 Further arguments to be passed to graphical functions.

Details

Eigenvalues larger than the upper bound of the monte carlo confidence limits are considered to be not explainable by an AR1 noise process.

Author(s)

Lukas Gudmundsson

See Also

MCSSA, par

[Package simsalabim version 0.1-1 Index]