plot.decompSSA(simsalabim)	R Documentation

Plot the SSA eigenspectrum

Description

Plots the SSA eigenspectrum.

Usage


## S3 method for class 'decompSSA':
plot(x, by = "freq", normalize = FALSE, asFreq = TRUE,
      ann = TRUE, log = "xy", ...)

Arguments

x 	 Object of class decompSSA

by 	 How eigenvalues shall be ordered. "rank" denotes ordering by decreasing order."freq" denotes ordering by their frequency.

normalize 	 If eigenvalues shall be normalized to their sum.

asFreq 	 Only is applied if by="freq". Determines whether the frequencies shall be expressed as true “frequency” or be inverted as “period”.

ann 	 Whether to do annotation or not.

log 	 Use logarithmic axes, usage like log in par.

... 	 Other arguments to be passed to graphical functions.

Details

The SSA eigenspectrum is build of the (potentially normalised) eigenvalues ordered either by the dominant frequency of the corresponding eigenvector or by their rank (decreasingly order).

Author(s)

Lukas Gudmundsson

See Also

decompSSA , decompSSAM

Examples


x <- sin(seq(0,10*pi,len=200))
x <- x + rnorm(x)/2
x.dc <- decompSSA(x,L=40)
plot(x.dc,by="rank",log="")
plot(x.dc,by="freq",log="y")

[Package simsalabim version 0.1-1 Index]