plot.sdTest(simsalabim)	R Documentation

Plots the results of the Shun and Duffy (1999) significance test.

Description

Draws the SSA - Spectrum (eigenvalues versus the dominant frequency of the eigenvectors) and draws the the fitted noise model as well as the confidence bounds of the model.

Usage


## S3 method for class 'sdTest':
plot(x, asFreq = TRUE, normalize = TRUE, lam.pch = 19,
      lam.col = "black", lam.cex = 1, sig.pch = 19,
      sig.col = "red", sig.cex = 1, conf.col = "lightgray",
      conf.border = "lightgray", noise.col = "black",
      noise.lwd = 1, ann = TRUE, legend = TRUE,
      axes = TRUE, ...)

Arguments

x 	 Output of sdTest

asFreq 	 If by="freq", whether frequency shall be plotted as frequency or as period.

normalize 	 Whether eigenvalues shall be normalised to their sum or not.

lam.pch 	 pch of the eigenvalues.

lam.col 	 col of the eigenvalues.

lam.cex 	 cex of the eigenvalues.

sig.pch 	 pch of the eigenvalues, that are significantly different from the noise spectrum.

sig.col 	 col of the eigenvalues, that are significantly different from the noise spectrum.

sig.cex 	 cex of the eigenvalues, that are significantly different from the noise spectrum.

conf.col 	 col of the confidence bounds of the noise spectrum, as defined in polygon.

conf.border 	 border of the confidence bounds of the noise spectrum, as defined in polygon.

noise.col 	 Colour of the line representing the fitted noise model.

noise.lwd 	 lwd of the line representing the fitted noise model.

ann 	 Whether to do annotations or not.

legend 	 Whether to add a legend or not

axes 	 Whether to add axes or not.

... 	 Further arguments to be passed to other graphic functions.

Author(s)

Lukas Gudmundsson

See Also

sdTest, par

[Package simsalabim version 0.1-1 Index]