sdTest(simsalabim)	R Documentation

Test SSA eigenspectrum against red noise

Description

Performs a test of an SSA eigenspectrum against a red noise hypothesis, following Shun and Duffy (1999).

Usage


sdTest(dSSA, ...)
## S3 method for class 'decompSSA':
sdTest(dSSA, conf = 0.95,...)

Arguments

dSSA 	 An object of class decompSSA

conf 	 Confidence limit

... 	 Further arguments to be passed to different methods.

Details

An implementation of the test introduced by Shun and Duffy (1999). The crucial step within the computations is a non linear fit by nls. In some cases nls throws an error which causes the test to fail. In such cases the SSA - spectrum does not have enough resemblance with the model to allow a sensible fit. Therefore, it can be assumed that the red noise null hypothesis is not appropriate anyway.

The confidence level conf can be specified either as number < 1 or as vector of length two with sum(conf)==1.

Value

An object of class sdTest

lambda 	 The eigenvalues, ordered by decreasing frequency.

freq 	 Dominant frequency of the eigenvectors, ordered by decreasing.

rank 	 Rank of the eigenvalues, ordered by decreasing frequency.

fittedNoise 	 Fitted values of the noise model, ordered by decreasing frequency.

upper 	 Upper confidence limit, ordered by decreasing frequency.

lower 	 Lower confidence limit, ordered by decreasing frequency.

conf 	 Vector of length two specifying the upper and lower confidence bounds of the test.

N 	 Length of the initial series

L 	 The embedding dimension

call 	 Call of the generating function

Warning

In some cases the function throws an error due to a failure of a non linear fitting procedure (see details above).

Author(s)

Lukas Gudmundsson

References

Shun, T. & Duffy, C. Low-frequency oscillations in precipitation, temperature, and runoff on a west facing mountain front: A hydrogeologic interpretation. Water Resources Research, 1999, 35, 191-201

See Also

decompSSA, decompSSAM, plot.sdTest,MCSSA

Examples


n <- arima.sim(n=200,alist(ar=0.4))
n <- as.vector(scale(n))
x <- sin(seq(0,15*pi,len=200))
xn <- x+n/2
xn.dc <- decompSSA(xn,L=40)
plot(sdt<-sdTest(xn.dc))


[Package simsalabim version 0.1-1 Index]