simsalabim-package(simsalabim)	R Documentation

A collection of methods for time series analysis and signal detection.

Description

This package implements Singular System Analysis and Phase Synchronization Indices.

Details

The functions bundled in this package focus on two objectives: The application of Singular Spectrum Analysis (SSA) and the computation of phase synchronization measures. The working horses of this package are the functions decompSSA and phaSyn. For both applications some statistical testing procedures are provided. SSA is implemented in a standard form and in a variant able to handle missing data points decompSSAM. Currently the entire package is implemented in R which leads to some performance issues.

Author(s)

Lukas Gudmundsson

Maintainer: Lukas Gudmundsson lukas.gudmundsson@geo.uio.no

References

Golyandina, N.; Nekrutkin, V. & Zhiglkilavskifi, A. Analysis of Time Series Structure: SSA and Related Techniques. CRC Press, 2001

Ghil, M.; Allen, M.; Dettinger, M.; Ide, K.; Kondrashov, D.; Mann, M.; Robertson, A.; Saunders, A.; Tian, Y.; Varadi, F. & others. Advanced spectral methods for climatic time series. Rev. Geophys, 2002, 40, 1003

Pikovsky, A.; Rosenblum, M. & Kurths, J. Synchronization: A Universal Concept in Nonlinear Sciences. (Cambridge Nonlinear Science Series) Cambridge University Press, 2003

[Package simsalabim version 0.1-1 Index]