synBoot(simsalabim)	R Documentation

Permutation test for phase synchronisation indices

Description

Computes a permutation test on phase synchronisation indices.

Usage


synBoot(x1,x2,singleSide=FALSE,method="MRL",M=100,R=100)

Arguments

x1, x2 	 Numeric vectors of the same length, interpreted as time series.

singleSide 	 If true only x1 is permutated.

method 	 How to determine the phase synchronisation index. "MRL": mean resultant length. "SH": a measure based on the Shannon - Entropy.

M 	 Number of bins in the histogram of the cyclic phase difference, needed to compute the Shannon - entropie based phase synchronisation index. This is only used if method="SH" and is ignored otherwise.

R 	 Number of replications for random permutation.

Details

Permutates the values of x1 and x2 R times and computes for every replication the phase synchronisation index. The resulting vector of surrogate phase synchronisation indices may be used to compute quantiles, that can be used for significance testing (see example below).

Value

A numeric vector of length R containing the Phase Synchronisation indices between both time series for each permutation.

Warning

May be computational expernsive

Author(s)

Lukas Gudmundsson

References

Allefeld, C. & Kurths, J. Testing for phase synchronization. International Journal of Bifurcation and Chaos, 2004, 14, 405-416

See Also

phaSyn

Examples



x1 <- sin(1:100)
x2 <- cos(1:100)
ps1 <- phaSyn(x1,x2)$rho
psp <- synBoot(x1,x2)
quantile(psp,0.95) < ps1 # Significance at the 95% confidence level?


[Package simsalabim version 0.1-1 Index]